Added functionality to load .tex file and edit section before saving.

+++ src/lgimapy/latex/document.py
@@ -11,6 +11,24 @@ from lgimapy.utils import mkdir, root, to_list
 # %%
 
 
+class OpenEditError(Exception):
+    """Raised when trying to save before finishing current edit."""
+
+    pass
+
+
+class NoEditFoundError(Exception):
+    """Raised when no current edit has been found."""
+
+    pass
+
+
+class KeywordNotFoundError(Exception):
+    """Raised when no current edit has been found."""
+
+    pass
+
+
 class Document:
     """
     Class for building a document in LaTeX.
@@ -37,9 +55,15 @@ class Document:
         * None: create in current directory.
         * Path: create at specified directory, create dir if required.
         * str: create `root/latex/{str}` directory, create dir if required.
+    fig_dir: bool, defautl=False
+        If True, create a directory named ``fig`` to store all
+        figures in the specified path.
+    load_tex: bool, default=False
+        If True, load a tex file with the specified filename
+        to edit.
     """
 
-    def __init__(self, fid, path=None, fig_dir=False):
+    def __init__(self, fid, path=None, fig_dir=False, load_tex=False):
         # Format fid and path.
         self.fid = fid[:-4] if fid.endswith(".tex") else fid
         if path is None:
@@ -60,11 +84,99 @@ class Document:
             self.fig_dir = self.path
 
         # Initialize document components.
-        self.body = "\n\n\\begin{document}\n\n"
-        self.add_preamble()
+        if load_tex:
+            self._loaded_file = True
+            self.preamble, self.body = self._load_tex()
+        else:
+            self._loaded_file = False
+            self.body = "\n\n\\begin{document}\n\n"
+            self.add_preamble()
         self.bibliography = ""
         self.appendix = ""
         self.background_image = ""
+        self._currently_open_edit = False
+
+    def _load_tex(self):
+        """
+        Load .tex file and separate preamble from body.
+
+        Returns
+        -------
+        preamble: str
+            Preamble of loaded .tex file.
+        body: str
+            Body of loaded .tex file.
+        """
+        # Load file.
+        fid = f"{self.path.joinpath(self.fid)}.tex"
+        with open(fid, "r") as f:
+            doc = f.read()
+
+        # Split into file parts.
+        ix_begin_body = doc.find("\\begin{document}")
+        ix_end_body = doc.find("\end{document}")
+        preamble = doc[:ix_begin_body]
+        body = doc[ix_begin_body:ix_end_body]
+        return preamble, body
+
+    def start_edit(self, keyword):
+        """
+        Begin editing saved document at keyword location.
+
+        Parameters
+        ----------
+        keyword: str
+            Keyword in .tex file to start edit.
+            In .tex file use ``%%\begin{<keyword>}`` to indicate
+            starting location for edit. Use ``%%\end{<keyword>}``
+            to indicate the end of the portion to be replaced.
+
+        See Also
+        --------
+        :meth:`Document.end_edit`
+        """
+        # Find editing location and raise error if start or end
+        # points are not declared.
+        start_phrase = f"%%\\begin{{{keyword}}}"
+        end_phrase = f"%%\\end{{{keyword}}}"
+        start_ix = self.body.find(start_phrase)
+        end_ix = self.body.find(end_phrase)
+        if start_ix == -1 or end_ix == -1:
+            msg = f"'{keyword}' not found as editing keyword in {self.fid}.tex"
+            raise KeywordNotFoundError(msg)
+
+        # Store keyword and sections of the document surrounding
+        # area to be edited to recombine later.
+        self._current_keyword = keyword
+        self._pre_edit = self.body[:start_ix]
+        self._post_edit = self.body[end_ix:]
+        self.body = ""
+
+        # Make safety flag to avoid saving file before edit is complete.
+        self._currently_open_edit = True
+
+    def end_edit(self):
+        """
+        Finish previously started edit.
+
+        See Also
+        --------
+        :meth:`Document.start_edit`
+        """
+        # Make sure an edit has been started.
+        if not self._currently_open_edit:
+            msg = "Document.start_edit() has not been started."
+            raise NoEditFoundError(msg)
+
+        # Combine pre- and post-edited sections of the .tex document
+        # with the newly created edited section.
+        start_phrase = f"%%\\begin{{{self._current_keyword}}}"
+        self.body = "\n\n".join(
+            [self._pre_edit, start_phrase, self.body, self._post_edit]
+        )
+
+        # Close edit so document can be safely saved.
+        self._currently_open_edit = False
 
     def add_preamble(
         self,
@@ -203,13 +315,41 @@ class Document:
         self.bibliography = ""
 
     def add_appendix(self):
+        """TODO"""
         self.appendix = ""
 
     def add_section(self, section):
+        """
+        Add a new section to :attr:`Document.body`.
+
+        Parameters
+        ----------
+        section: str
+            Title of new section.
+        """
         self.body = "\n\n".join([self.body, f"\section{{{section}}}"])
 
     def add_paragraph(self, paragraph):
-        self.body = "\n\n".join([self.body, paragraph, "\\bigskip"])
+        """
+        Add text as a new paragraph to :attr:`Document.body`.
+
+        Parameters
+        ----------
+        paragraph: str
+            Individual paragraph of text.
+        """
+        self.body = "\n".join([self.body, "\n", paragraph, "\\bigskip"])
+
+    def add_text(self, text):
+        """
+        Add text to :attr:`Document.body`.
+
+        Parameters
+        ----------
+        text: str
+            Text to add to :attr:`Document.body`.
+        """
+        self.body = "\n\n".join([self.body, text])
 
     def add_background_image(
         self,
@@ -332,8 +472,11 @@ class Document:
             ]
         )
 
-    def write_tex(self):
+    def save_tex(self):
         """Save .tex file of current document."""
+        if self._currently_open_edit:
+            raise OpenEditError("Cannot save before current edit is completed.")
+
         # Clean file path and move to proper directory.
         if self.path:
             os.chdir(self.path)
@@ -354,6 +497,9 @@ class Document:
         save_tex: bool, default=False
             If True, save `.tex` as well as `.pdf`.
         """
+        if self._currently_open_edit:
+            raise OpenEditError("Cannot save before current edit is completed.")
+
         # Clean file path and move to proper directory.
         if self.path:
             os.chdir(self.path)
@@ -375,12 +521,11 @@ class Document:
                 stderr=subprocess.STDOUT,
                 stdout=subprocess.DEVNULL,
             )
-        if not save_tex:
-            os.remove(f"{fid}.tex")
 
+        # Clean up temporary files.
+        if not save_tex and not self._loaded_file:
+            print("deleted")
+            os.remove(f"{fid}.tex")
         os.remove(f"{fid}.aux")
         os.remove(f"{fid}.log")
         os.remove(f"{fid}.out")
-
-
-# %%
